freezed-requirements.txt
# This file may be used to create an environment using:
# $ `conda create --name <env> --file <this file>`
# platform: osx-64
astra-toolbox=1.8.3=py36hb466ac3_0
attrs=21.2.0=pyhd3eb1b0_0
blas=1.0=mkl
ca-certificates=2021.10.26=hecd8cb5_2
certifi=2021.5.30=py36hecd8cb5_0
cycler=0.10.0=py36hecd8cb5_0
freetype=2.10.4=ha233b18_0
future=0.18.2=pypi_0
h5py=2.10.0=py36h3134771_0
hdf5=1.10.4=hfa1e0ec_0
importlib-metadata=3.10.0=py36hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.3.0=hecd8cb5_3375
jpeg=9b=he5867d9_2
kiwisolver=1.3.1=py36h23ab428_0
lcms2=2.12=hf1fd2bf_0
libastra=1.8.3=he374074_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=h87d7836_0
libwebp-base=1.2.0=h9ed2024_0
lz4-c=1.9.3=h23ab428_0
matplotlib=3.1.1=py36h54f8f79_0
meson=0.59.0=pypi_0
mkl=2019.4=233
mkl-service=2.3.0=py36h9ed2024_0
more-itertools=8.8.0=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
numexpr=2.7.3=pypi_0
numpy=1.13.3=pypi_0
odl=1.0.0.dev0=dev_0
olefile=0.46=py36_0
openjpeg=2.3.0=hb95cd4c_1
openssl=1.1.1l=h9ed2024_0
packaging=21.0=pyhd3eb1b0_0
pandas=0.25.3=py36h0a44026_0
pillow=8.3.1=py36ha4cf6ea_0
pip=21.2.2=py36hecd8cb5_0
pluggy=0.13.1=py36hecd8cb5_0
proximal=0.1.7=pypi_0
py=1.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pytest=6.2.4=py36hecd8cb5_2
python=3.6.13=h88f2d9e_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=pypi_0
readline=8.1=h9ed2024_0
scipy=1.5.4=pypi_0
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py36hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hce871da_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py36h9ed2024_0
typing_extensions=3.10.0.0=pyh06a4308_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
